refactor: move inner classes to top level
Move the gRPC-related inner classes from AbstractResultSet to
top-level classes, so they are easier to modify and maintain.

This change only contains modifications that are needed to move
these inner classes. There are no functional changes.
