+++ google-cloud-spanner/src/test/java/com/google/cloud/spanner/ReadFormatTestRunner.java
@@ -104,9 +104,9 @@ public class ReadFormatTestRunner extends ParentRunner<JSONObject> {
   }
 
   private static class TestCaseRunner {
-    private AbstractResultSet.GrpcResultSet resultSet;
+    private GrpcResultSet resultSet;
     private SpannerRpc.ResultStreamConsumer consumer;
-    private AbstractResultSet.GrpcStreamIterator stream;
+    private GrpcStreamIterator stream;
     private JSONObject testCase;
 
     TestCaseRunner(JSONObject testCase) {
@@ -114,7 +114,7 @@ public class ReadFormatTestRunner extends ParentRunner<JSONObject> {
     }
 
     private void run() throws Exception {
-      stream = new AbstractResultSet.GrpcStreamIterator(10);
+      stream = new GrpcStreamIterator(10);
       stream.setCall(
           new SpannerRpc.StreamingCall() {
             @Override
@@ -130,7 +130,7 @@ public class ReadFormatTestRunner extends ParentRunner<JSONObject> {
           },
           false);
       consumer = stream.consumer();
-      resultSet = new AbstractResultSet.GrpcResultSet(stream, new NoOpListener());
+      resultSet = new GrpcResultSet(stream, new NoOpListener());
 
       JSONArray chunks = testCase.getJSONArray("chunks");
       JSONObject expectedResult = testCase.getJSONObject("result");
@@ -143,8 +143,7 @@ public class ReadFormatTestRunner extends ParentRunner<JSONObject> {
       assertResultSet(resultSet, expectedResult.getJSONArray("value"));
     }
 
-    private void assertResultSet(AbstractResultSet.GrpcResultSet actual, JSONArray expected)
-        throws Exception {
+    private void assertResultSet(GrpcResultSet actual, JSONArray expected) throws Exception {
       int i = 0;
       while (actual.next()) {
         Struct actualRow = actual.getCurrentRowAsStruct();
