+++ google-cloud-spanner/src/test/java/com/google/cloud/spanner/ResultSetsHelper.java
@@ -17,7 +17,6 @@
 package com.google.cloud.spanner;
 
 import com.google.cloud.spanner.AbstractResultSet.CloseableIterator;
-import com.google.cloud.spanner.AbstractResultSet.GrpcResultSet;
 import com.google.cloud.spanner.AbstractResultSet.Listener;
 import com.google.protobuf.ListValue;
 import com.google.spanner.v1.PartialResultSet;
