+++ google-cloud-spanner/src/test/java/com/google/cloud/spanner/ResumableStreamIteratorTest.java
@@ -23,7 +23,6 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 
 import com.google.api.client.util.BackOff;
-import com.google.cloud.spanner.AbstractResultSet.ResumableStreamIterator;
 import com.google.cloud.spanner.v1.stub.SpannerStubSettings;
 import com.google.common.collect.AbstractIterator;
 import com.google.common.collect.Lists;
@@ -53,7 +52,7 @@ import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 import org.mockito.Mockito;
 
-/** Unit tests for {@link AbstractResultSet.ResumableStreamIterator}. */
+/** Unit tests for {@link ResumableStreamIterator}. */
 @RunWith(JUnit4.class)
 public class ResumableStreamIteratorTest {
   interface Starter {
@@ -128,7 +127,7 @@ public class ResumableStreamIteratorTest {
   }
 
   Starter starter = Mockito.mock(Starter.class);
-  AbstractResultSet.ResumableStreamIterator resumableStreamIterator;
+  ResumableStreamIterator resumableStreamIterator;
 
   @Before
   public void setUp() {
@@ -137,7 +136,7 @@ public class ResumableStreamIteratorTest {
 
   private void initWithLimit(int maxBufferSize) {
     resumableStreamIterator =
-        new AbstractResultSet.ResumableStreamIterator(
+        new ResumableStreamIterator(
             maxBufferSize,
             "",
             null,
